+++ src/components/widgets/Checkbox.svelte
@@ -40,16 +40,36 @@
 <style>
     [type='checkbox'] {
         appearance: none;
-        border: solid var(--wordplay-border-color) var(--wordplay-border-width);
+        border: var(--wordplay-border-width) solid var(--wordplay-border-color);
         width: 1rem;
         height: 1rem;
         min-width: 1rem;
         min-height: 1rem;
         cursor: pointer;
+        box-shadow: var(--wordplay-border-width) var(--wordplay-border-width) 0
+            var(--wordplay-border-color);
+        transition:
+            transform calc(var(--animation-factor) * 100ms),
+            box-shadow calc(var(--animation-factor) * 100ms);
+    }
+
+    [type='checkbox']:hover {
+        transform: translate(-1px, -1px);
+    }
+
+    [type='checkbox']:focus {
+        outline: none;
+        box-shadow: var(--wordplay-border-width) var(--wordplay-border-width) 0
+            var(--wordplay-focus-color);
     }
 
     [type='checkbox']:checked {
         background: var(--wordplay-foreground);
+        box-shadow: 0 0 0 var(--wordplay-border-color);
+        transform: translate(
+            var(--wordplay-border-width),
+            var(--wordplay-border-width)
+        );
     }
 
     input:indeterminate {