+++ CHANGELOG.md
@@ -10,6 +10,7 @@ These notes are publicly posted in [production](https://wordplay.dev/updates), s
 
 - We added a Portugese (Portugal) machine translation! Time to get to work to make it good.
 - We added `/design` page so that anyone working on design or development for the platform have an easier guide on what colors, fonts, spacing, and components to use (#935).
+- We made the user interface components more consistent, usable, and visible.
 
 ## Changed